requirements.txt
Flask==2.3.0
flask-mysql==1.5.1
flask-cors==3.0.10
mysql-connector-python==8.1.0
